Made the True64 compiler happy, fixed a 'LL' suffix bug.
git-svn-id: http://svn.boost.org/svn/boost/trunk@28695 b8fc166d-592f-0410-95f2-cb63ce0dd405

+++ boost/libs/wave/test/testwave/testwave.cpp
@@ -77,9 +77,10 @@ main(int argc, char *argv[])
         cmdline_options.add(desc_cmdline).add(app.common_options());
 
     // parse command line
+        // (the (int) cast is to make the True64 compiler happy)
         using namespace boost::program_options::command_line_style;
         po::parsed_options opts(po::parse_command_line(argc, argv, 
-            cmdline_options, unix_style, cmd_line_utils::at_option_parser));
+            cmdline_options, (int)unix_style, cmd_line_utils::at_option_parser));
         
         po::store(opts, vm);
         po::notify(vm);
